training/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
bleach==6.0.0
boto3==1.28.0
botocore==1.31.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-log==0.4.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.2
dataclasses-json==0.5.9
deprecation==2.1.0
dnspython==2.3.0
docutils==0.20.1
docx2txt==0.8
dotty-dict==1.3.1
et-xmlfile==1.1.0
fastapi==0.99.1
filetype==1.2.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
gotrue==1.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httptools==0.6.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.7.0
invoke==1.7.3
jaraco.classes==3.2.3
jmespath==1.0.1
joblib==1.3.1
keyring==24.2.0
kombu==5.3.1
langchain==0.0.224
langchainplus-sdk==0.0.20
loguru==0.7.0
lxml==4.9.3
Markdown==3.4.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
more-itertools==9.1.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.0
olefile==0.46
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==10.0.0
pinecone-client==2.2.2
pkginfo==1.9.6
postgrest==0.10.6
prompt-toolkit==3.0.39
pycparser==2.21
pycurl==7.45.2
pydantic==1.10.11
Pygments==2.15.1
pypandoc==1.11
pypdf==3.12.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-gitlab==3.15.0
python-magic==0.4.27
python-pptx==0.6.21
python-semantic-release==7.33.2
pytz==2023.3
PyYAML==6.0
readme-renderer==40.0
realtime==1.0.0
redis==4.6.0
regex==2023.6.3
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
s3transfer==0.6.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
storage3==0.5.2
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.2.2
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.4.0
tomlkit==0.11.8
tqdm==4.65.0
twine==3.8.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unstructured==0.8.1
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==10.4
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
